applied a patch to fix segmentation fault on 32-bit big-endian platforms by Niko Tyni (#45)
https://salsa.debian.org/perl-team/modules/packages/libdbd-sqlite3-perl/blob/master/debian/patches/0001-Fix-SQLITE_DBCONFIG_DEFENSIVE-parameter-types.patch

+++ dbdimp.c
@@ -464,7 +464,7 @@ sqlite_db_login6(SV *dbh, imp_dbh_t *imp_dbh, char *dbname, char *user, char *pa
         if (hv_exists(hv, "sqlite_defensive", 16)) {
             val = hv_fetch(hv, "sqlite_defensive", 16, 0);
             if (val && SvIOK(*val)) {
-                sqlite3_db_config(imp_dbh->db, SQLITE_DBCONFIG_DEFENSIVE, SvIV(*val), 0);
+                sqlite3_db_config(imp_dbh->db, SQLITE_DBCONFIG_DEFENSIVE, (int)SvIV(*val), 0);
             }
         }
     }
